requirements-hf.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
pandas==2.1.0
agno==0.1.0
